LPD-26752 Move setup into a new file
